+++ examples/django/components/django/component.py
@@ -1,6 +1,6 @@
 from batou.component import Component
-from batou.lib.python import VirtualEnv, Package
-from batou.lib.file import SyncDirectory, File
+from batou.lib.python import AppEnv
+from batou.lib.file import SyncDirectory
 from batou.lib.supervisor import Program
 from batou.utils import Address
 
@@ -9,10 +9,7 @@ class Django(Component):
 
     def configure(self):
         self.address = Address(self.host.fqdn, '8081')
-        venv = VirtualEnv('2.7')
-        self += venv
-
-        venv += Package('django', version='1.5.4')
+        self += AppEnv('3.6')
 
         self += SyncDirectory('mysite', source='mysite')