+++ src/batou/component.py
@@ -210,7 +210,8 @@ class Component(object):
             # special attribute handling to catch up.
             setattr(self, k, v)
 
-        self.configure()
+        with self.chdir(self.defdir):
+            self.configure()
         self += self._get_platform()
         self._platform_component = self._
         self.__setup_event_handlers__()
@@ -305,7 +306,7 @@ class Component(object):
                             # XXX?!?!?
                             raise batou.UpdateNeeded()
                         raise
-            except batou.UpdateNeeded:
+            except AssertionError:
                 self.__trigger_event__(
                     'before-update', predict_only=predict_only)
                 output.annotate(self._breadcrumbs)